firegloss_backend/requirements.txt
firebase-admin==6.4.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0